requirements.txt
altgraph==0.17.4
argon2-cffi==20.1.0
async-generator==1.10
attrs==23.2.0
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==3.2.1
blinker==1.8.2
bs4==0.0.2
certifi==2024.7.4
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
exceptiongroup==1.2.2
Faker==5.8.0
flask==3.0.3
flask-restx==1.3.0
h11==0.14.0
idna==3.7
imageio==2.33.1
imageio-ffmpeg==0.4.9
importlib-metadata==8.2.0
ipykernel==5.4.3
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==2.2.0
jedi==0.18.0
jinja2==3.1.4
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.7.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab-pygments==0.1.2
lxml==4.6.2
MarkupSafe==2.1.5
mistune==0.8.4
moviepy==1.0.3
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
notebook==6.2.0
numpy==1.24.4
outcome==1.3.0.post0
packaging==20.8
pandocfilters==1.4.3
parso==0.8.1
pefile==2023.2.7
pickleshare==0.7.5
Pillow==10.1.0
proglog==0.1.10
prometheus-client==0.9.0
prompt-toolkit==3.0.10
pycparser==2.22
Pygments==2.7.4
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
pyparsing==2.4.7
pyperclip==1.9.0
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-dotenv==1.0.1
pytest==8.0.0
pytz==2024.2
pywin32==300
pywin32-ctypes==0.2.2
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
requests==2.31.0
selenium==4.23.1
selenium-stealth==1.0.6
Send2Trash==1.5.0
six==1.15.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SpeechRecognition==3.10.1
terminado==0.9.2
testpath==0.4.4
text-unidecode==1.3
tornado==6.1
tqdm==4.66.1
traitlets==5.0.5
trio==0.26.0
trio-websocket==0.11.1
typing-extensions==4.12.2
unicodecsv==0.14.1
urllib3==2.2.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.8.0
werkzeug==3.0.3
wsproto==1.2.0
zipp==3.19.2
